Extract QueryCompiler to its own file
