+++ core/src/main/scala/playground/QueryCompilerVisitor.scala
@@ -1,9 +1,7 @@
 package playground
 
-import cats.Apply
 import cats.Id
 import cats.data.Ior
-import cats.data.IorNec
 import cats.data.NonEmptyChain
 import cats.data.NonEmptyList
 import cats.implicits._
@@ -52,57 +50,9 @@ import java.util.UUID
 
 import types._
 import util.chaining._
-import QueryCompiler.WAST
-
-trait QueryCompiler[A] {
-  final def emap[B](f: A => QueryCompiler.Result[B]): QueryCompiler[B] =
-    ast => compile(ast).flatMap(f)
-
-  def compile(ast: WAST): QueryCompiler.Result[A]
-
-}
-
-object QueryCompiler {
-  type Result[+A] = IorNec[CompilationError, A]
-
-  implicit val apply: Apply[QueryCompiler] =
-    new Apply[QueryCompiler] {
-      def map[A, B](fa: QueryCompiler[A])(f: A => B): QueryCompiler[B] = fa.compile(_).map(f)
-
-      def ap[A, B](ff: QueryCompiler[A => B])(fa: QueryCompiler[A]): QueryCompiler[B] =
-        wast => (ff.compile(wast), fa.compile(wast)).parMapN((a, b) => a(b))
-    }
-
-  type WAST = WithSource[InputNode[WithSource]]
-
-  val pos: QueryCompiler[SourceRange] = _.range.rightIor
-  val unit: QueryCompiler[Unit] = _ => ().rightIor
-
-  def typeCheck[A](
-    expected: NodeKind
-  )(
-    f: PartialFunction[InputNode[WithSource], A]
-  ): QueryCompiler[WithSource[A]] =
-    ast =>
-      ast
-        .traverse(f.lift)
-        .toRightIor(
-          NonEmptyChain(
-            CompilationError.error(
-              TypeMismatch(
-                expected,
-                ast.value.kind,
-              ),
-              ast.range,
-            )
-          )
-        )
-
-}
 
 object QueryCompilerVisitor {
   val full = new TransitiveCompiler(AddDynamicRefinements) andThen QueryCompilerVisitorInternal
-
 }
 
 object QueryCompilerVisitorInternal extends SchemaVisitor[QueryCompiler] {
@@ -431,8 +381,7 @@ object QueryCompilerVisitorInternal extends SchemaVisitor[QueryCompiler] {
     it.value.compile(_)
   }
 
-  val stringLiteral =
-    QueryCompiler.typeCheck(NodeKind.StringLiteral) { case StringLiteral(s) => s }
+  val stringLiteral = QueryCompiler.typeCheck(NodeKind.StringLiteral) { case StringLiteral(s) => s }
 
   val document: QueryCompiler[Document] =
     _.value match {
